[first-query] dir and file struct refactor
