+++ Cargo.toml
@@ -24,3 +24,16 @@ tokio = { version = "1.25.0", features = ["macros", "rt-multi-thread"] }
 
 [dev-dependencies]
 reqwest = "0.11.14"
+
+# Using table-like toml syntax to avoid a super-long line!
+[dependencies.sqlx]
+version = "0.6"
+default-features = false
+features = [
+    "runtime-tokio-rustls",
+    "macros",
+    "postgres",
+    "uuid",
+    "chrono",
+    "migrate"
+]
