+++ src/main.rs
@@ -1,7 +1,7 @@
 //! main.rs
 
 use std::net::TcpListener;
-use zero2prod::run;
+use zero2prod::startup::run;
 
 #[tokio::main]
 async fn main() -> std::io::Result<()> {
