+++ src/routes/subscriptions.rs
@@ -1,0 +1,11 @@
+use actix_web::{HttpResponse, web};
+
+#[derive(serde::Deserialize)]
+pub struct FormData {
+    email: String,
+    name: String,
+}
+
+pub async fn subscribe(_form: web::Form<FormData>) -> HttpResponse {
+    HttpResponse::Ok().finish()
+}
